Fixed the AsciiAssetGui to properly ferry UI calls to the UI thread when handling asset open/close. Right now, asset open/close is only happening on the UI thread, so we weren't noticing any problems, however, this would fail as soon as tests start opening / closing assets.
Fixed the ShellControl to be disablable, so that when the asset is closed, we don't try to send anything, nor do we let the user even try to enter commands.

Fixed the ShellControl to use a System.Windows.Forms.Timer object for pumping data back to the UI. This is better than threads because there is no thread-thunking needed (the UI thread reads from our thread-safe RingBuffer) and doesn't require that we have all of the resources fired up to manage N threads.