Finished implementing Add. Our design is to pre-emptively split interior nodes that are full as we're recursing down through them to find the leaf node that needs inserting. Consider this scenario:
* Traverse Root
* Traverse Root.IntChild5
* Traverse Root.IntChild5.IntChild13
* Traverse Root.IntChild5.IntChild13.IntChild20
* Find leaf at Root.IntChild5.IntChild13.IntChild20.Leaf20

Its possible that every node in that chain is full. We'll go to split Leaf20 and insert the new sibling into IntChild20, which is full, so we split IntChild20 etc etc upward until we've split everybody. But doing so like that means that we have to hold locks for that entire chain all the way back up.

Instead, we pre-emptively split. We allow inserts to fill a node, but if we're performing another insert and find that an intermediate node is full, we'll pre-emptively split that node. This means that we can recurse by holding onto only a node and its parent, and we're guaranteed that at any step, the parent isn't full.

-----

I Changed the design of InteriorNode so that when it is inserting a node, it takes the insert key explicitly instead of taking it from the first key in the child. This is necessary for having InteriorNodes that are children of InteriorNodes, since the key from a parent is strictly less or equal to all keys in all children on that side of the branch, not the first key, since child InteriorNodes have values that are less than their first key.