Fixed CharacterIndexToChunkCoord to only do one thing - figure out where a character index is in the line model, and tell me the number of pixels up to and including the character. This makes it simple and effective, and there's no question what to do about the space before the first character (which is simple - pixel 0).
I changed PixelToNearestText in the same way - instead of trying to indicate (inconsistently) the beginning of some character or the end of another character, we indicate them separately. First, we tell them the nearest character, which is useful in its own regard, and then we tell them whether the pixel is before or after the midpoint of the character, for positioning the caret.

For the time being, I've disabled the estimation logic just to simplify it while working on making it work. 

AS this stands, it works and correctly positions the caret.