Fixed properties block handling.
The presence of an empty property block is different than no property block at all. If the node is a change node, and it has no property block, then any previous properties persist. If the node is a change node, and it has an empty property block, then all previous properties are wiped.