Fixed a few bugs. Some boms are prefixes of others - so we need to keep looking until we whittle down all candidates into a few winners, and then pick the winner with the longest prefix match.
Since we may read a different number of bytes than the bom, we need to explicit store how many bytes we read and how many went into the bom, and return the extras in our unusedBytes variable.

Refactored the ReadDoc method to extract the portion that looks for lines in a chunk of bytes.